+++ packages/@expo/config/src/buildCacheProvider.ts
@@ -70,18 +70,29 @@ export type BuildCacheProvider<T = any> = {
 };
 
 export type BuildCacheProviderPlugin<T = any> = {
-  resolveBuildCache(props: ResolveRemoteBuildCacheProps, options: T): Promise<string | null>;
-  uploadBuildCache(props: UploadRemoteBuildCacheProps, options: T): Promise<string | null>;
   calculateFingerprintHash?: (
     props: CalculateFingerprintHashProps,
     options: T
   ) => Promise<string | null>;
-  /**
-   * @deprecated Use `resolveBuildCache` instead.
-   */
-  resolveRemoteBuildCache(props: ResolveRemoteBuildCacheProps, options: T): Promise<string | null>;
-  /**
-   * @deprecated Use `uploadBuildCache` instead.
-   */
-  uploadRemoteBuildCache(props: UploadRemoteBuildCacheProps, options: T): Promise<string | null>;
-};
+} & (
+  | {
+      resolveBuildCache(props: ResolveBuildCacheProps, options: T): Promise<string | null>;
+      uploadBuildCache(props: UploadBuildCacheProps, options: T): Promise<string | null>;
+    }
+  | {
+      /**
+       * @deprecated Use `resolveBuildCache` instead.
+       */
+      resolveRemoteBuildCache: (
+        props: ResolveRemoteBuildCacheProps,
+        options: T
+      ) => Promise<string | null>;
+      /**
+       * @deprecated Use `uploadBuildCache` instead.
+       */
+      uploadRemoteBuildCache: (
+        props: UploadRemoteBuildCacheProps,
+        options: T
+      ) => Promise<string | null>;
+    }
+);