+++ packages/expo-background-task/src/BackgroundTask.ts
@@ -123,20 +123,20 @@ export async function unregisterTaskAsync(taskName: string): Promise<void> {
 /**
  * When in debug mode this function will trigger running the background tasks.
  * This function will only work for apps built in debug mode.
- * @todo(chrfalch): When we have a usable devtools plugin we can enable this function.
+ * This method is only available in development mode. It will not work in production builds.
  * @returns A promise which fulfils when the task is triggered.
  */
-// export async function triggerTaskWorkerForTestingAsync(): Promise<boolean> {
-//   if (__DEV__) {
-//     if (!ExpoBackgroundTaskModule.triggerTaskWorkerForTestingAsync) {
-//       throw new UnavailabilityError('BackgroundTask', 'triggerTaskWorkerForTestingAsync');
-//     }
-//     console.log('Calling triggerTaskWorkerForTestingAsync');
-//     return await ExpoBackgroundTaskModule.triggerTaskWorkerForTestingAsync();
-//   } else {
-//     return Promise.resolve(false);
-//   }
-// }
+export async function triggerTaskWorkerForTestingAsync(): Promise<boolean> {
+  if (__DEV__) {
+    if (!ExpoBackgroundTaskModule.triggerTaskWorkerForTestingAsync) {
+      throw new UnavailabilityError('BackgroundTask', 'triggerTaskWorkerForTestingAsync');
+    }
+    console.log('Calling triggerTaskWorkerForTestingAsync');
+    return await ExpoBackgroundTaskModule.triggerTaskWorkerForTestingAsync();
+  } else {
+    return Promise.resolve(false);
+  }
+}
 
 // Export types
 export {