+++ packages/expo-maps/ios/MapRecords.swift
@@ -84,6 +84,56 @@ struct ExpoAppleMapPolyline: Record, Identifiable {
   }
 }
 
+struct Circle: Record, Identifiable {
+  @Field var id: String = UUID().uuidString
+
+  @Field var center: Coordinate
+  @Field var radius: Double
+  @Field var lineColor: Color?
+  @Field var lineWidth: Double?
+  @Field var color: Color = .green
+
+  var clLocationCoordinate2D: CLLocationCoordinate2D {
+    CLLocationCoordinate2D(
+      latitude: center.latitude,
+      longitude: center.longitude
+    )
+  }
+
+  var mkPlacemark: MKPlacemark {
+    MKPlacemark(coordinate: clLocationCoordinate2D)
+  }
+
+  var mapItem: MKMapItem {
+    MKMapItem(placemark: mkPlacemark)
+  }
+}
+
+struct Polygon: Record, Identifiable {
+  @Field var id: String = UUID().uuidString
+
+  @Field var coordinates: [Coordinate]
+  @Field var color: Color = .blue
+  @Field var lineColor: Color?
+  @Field var lineWidth: Double?
+
+  var clLocationCoordinates2D: [CLLocationCoordinate2D] {
+    return coordinates.map {
+      CLLocationCoordinate2D(latitude: $0.latitude, longitude: $0.longitude)
+    }
+  }
+
+  var mkPlacemark: MKPlacemark {
+    MKPlacemark(
+      coordinate: clLocationCoordinates2D.first ?? CLLocationCoordinate2D(latitude: 0, longitude: 0)
+    )
+  }
+
+  var mapItem: MKMapItem {
+    MKMapItem(placemark: mkPlacemark)
+  }
+}
+
 struct MapUISettings: Record {
   @Field var compassEnabled: Bool = true
   @Field var myLocationButtonEnabled: Bool = true
@@ -92,6 +142,7 @@ struct MapUISettings: Record {
 }
 
 struct MapProperties: Record {
+  @Field var isMyLocationEnabled: Bool = false
   @Field var mapType: MapType = .standard
   @Field var isTrafficEnabled: Bool = false
   @Field var selectionEnabled: Bool = true