+++ packages/expo-maps/src/google/GoogleMaps.types.ts
@@ -9,7 +9,7 @@ import { CameraPosition, Coordinates } from '../shared.types';
  */
 export type GoogleMapsMarker = {
   /**
-   * The unique identifier for the marker. This can be used to e.g. identify the clicked marker in the `onMarkerClick` event.
+   * The unique identifier for the marker. This can be used to identify the clicked marker in the `onMarkerClick` event.
    */
   id?: string;
 
@@ -49,7 +49,7 @@ export type GoogleMapsMarker = {
  */
 export type GoogleMapsPolyline = {
   /**
-   * The unique identifier for the polyline. This can be used to e.g. identify the clicked polyline in the `onPolylineClick` event.
+   * The unique identifier for the polyline. This can be used to identify the clicked polyline in the `onPolylineClick` event.
    */
   id?: string;
 
@@ -72,6 +72,71 @@ export type GoogleMapsPolyline = {
    * Whether the polyline is geodesic.
    */
   geodesic?: boolean;
+};
+
+/**
+ * @platform android
+ */
+export type GoogleMapsCircle = {
+  /**
+   * The unique identifier for the circle. This can be used to identify the clicked circle in the `onCircleClick` event.
+   */
+  id?: string;
+
+  /**
+   * The coordinates of the circle.
+   */
+  center: Coordinates;
+
+  /**
+   * The radius of the circle.
+   */
+  radius: number;
+
+  /**
+   * The color of the circle.
+   */
+  color?: ProcessedColorValue | string;
+
+  /**
+   * The color of the circle line.
+   */
+  lineColor?: ProcessedColorValue | string;
+
+  /**
+   * The width of the circle line.
+   */
+  lineWidth?: number;
+};
+
+/**
+ * @platform android
+ */
+export type GoogleMapsPolygon = {
+  /**
+   * The unique identifier for the polygon. This can be used to identify the clicked polygon in the `onPolygonClick` event.
+   */
+  id?: string;
+
+  /**
+   * The coordinates of the circle.
+   */
+  coordinates: Coordinates[];
+
+  /**
+   * The color of the polygon.
+   */
+  color?: ProcessedColorValue | string;
+
+  /**
+   * The width of the polygon.
+   */
+  lineWidth?: number;
+
+  /**
+   * The color of the polygon.
+   */
+  lineColor?: ProcessedColorValue | string;
 };
 
 /**
@@ -255,6 +320,16 @@ export type GoogleMapsViewProps = {
   polylines?: GoogleMapsPolyline[];
 
   /**
+   * The array of polygons to display on the map.
+   */
+  polygons?: GoogleMapsPolygon[];
+
+  /**
+   * The array of circles to display on the map.
+   */
+  circles?: GoogleMapsCircle[];
+
+  /**
    * The `MapUiSettings` to be used for UI-specific settings on the map.
    */
   uiSettings?: GoogleMapsUISettings;
@@ -304,6 +379,16 @@ export type GoogleMapsViewProps = {
    * Lambda invoked when the polyline is clicked.
    */
   onPolylineClick?: (event: GoogleMapsPolyline) => void;
+
+  /**
+   * Lambda invoked when the polygon is clicked.
+   */
+  onPolygonClick?: (event: GoogleMapsPolygon) => void;
+
+  /**
+   * Lambda invoked when the circle is clicked.
+   */
+  onCircleClick?: (event: GoogleMapsCircle) => void;
 
   /**
    * Lambda invoked when the map was moved by the user.