+++ packages/jest-expo/src/preset/setup.js
@@ -4,13 +4,16 @@
  */
 'use strict';
 
+import merge from 'lodash/merge';
+
 const findUp = require('find-up');
 const path = require('path');
 const mockNativeModules = require('react-native/Libraries/BatchedBridge/NativeModules').default;
 const stackTrace = require('stacktrace-js');
 
-const publicExpoModules = require('./expoModules');
-const internalExpoModules = require('./internalExpoModules');
+const publicExpoModules = require('./moduleMocks/expoModules');
+const internalExpoModules = require('./moduleMocks/internalExpoModules');
+const thirdPartyModules = require('./moduleMocks/thirdPartyModules');
 
 // window isn't defined as of react-native 0.45+ it seems
 if (typeof window !== 'object') {
@@ -51,10 +54,7 @@ Object.defineProperty(mockNativeModules, 'LinkingManager', {
   get: () => mockNativeModules.Linking,
 });
 
-const expoModules = {
-  ...publicExpoModules,
-  ...internalExpoModules,
-};
+const expoModules = merge(publicExpoModules, merge(thirdPartyModules, internalExpoModules));
 
 // Mock the experience URL in development mode for asset setup
 expoModules.NativeUnimoduleProxy.modulesConstants.mockDefinition.ExponentConstants.experienceUrl.mock =