+++ packages/expo/CHANGELOG.md
@@ -10,6 +10,12 @@
 
 ### 💡 Others
 
+## 53.0.9 — 2025-05-08
+
+### 🐛 Bug fixes
+
+- respect `react-native` type export conditions ([#36728](https://github.com/expo/expo/pull/36728) by [@vonovak](https://github.com/vonovak))
+
 ## 53.0.8 — 2025-05-06
 
 _This version does not introduce any user-facing changes._